job-manager: do not send purged events
Problem: If a job has been purged from the job-manager, events
in the journal may still exist for those purged jobs.  They can
be sent to journal requestors.

Solution: When a new request is made for the journal, check that
the events in the journal are still valid before sending them.

Fixes #4331