testsuite: workaround issue4333 in tests
Problem: As detailed in issue #4333, when the kvs module is reloaded,
the kvs-watch module has to be reloaded at the moment.  This
was not done in issues/t4222-kvs-assume-empty-dir.sh.  This lead
to potential problems in follow on tests.

Solution: Remove / reload the kvs-watch module as well as the
kvs module in t4222-kvs-assume-empty-dir.sh.