python: fix incorrect variable in pid resolver error message
Problem: An incorrect variable name is used in an error message
from the pid URI resolver when a flux-broker with no children is
detected. Therefore, instead of sensible error message, users see an
invalid name error.

Fix the error message to use the correct variable name so that a
sensible error is emitted.