t2802-uri-cmd.t: add test for PID resolver on broker with no child
Problem: The URI resolver tests do not exercise an error condition
where the PID target is a flux-broker with no children. This let a
bug creep through undetected.

Add a test to ensure that the PID URI resolver on a flux-broker
process with no children raises an error such as:

 flux-uri: ERROR: PID 28628 is a flux-broker and no child found