kvs: support commit syncing
Problem: It would be convenient if a kvs commit could also
flush and checkpoint right after the commit completes.  This would
ensure committed data is on the backing store and
will be recovered via checkpoint if necessary.

Solution: Add a FLUX_KVS_SYNC flag to flux_kvs_commit().  Update
kvs to perform a "content.flush" and "checkpoint" after a commit
completes if the flag is set.

Fixes #4308