flux-keygen: add --name option
Problem: flux-keygen always uses the hostname for the certificate
'name' metadata, but this name is logged when a peer tries to authenticate.
In a system instance, these log messages can be confusing since the peer
is likely not the host that generated the certificate.

Add -n,--name option which allows the certificate name to be set on the
command line.