libzmqutil: change auth log format used by broker
Problem: the broker logs a terse message displaying the certificate
name without adornment when a peer attempts to authenticate, but the
certificate name defaults to the name of the host that generated it,
which could easily be confused with the peer hostname.

Print cert-name=NAME instead of just the bare name to hopefully make
things a bit clearer.

Fixes #4341