job-manager: post submit event
Problem: job-ingest posts the submit event, but every other job
event is posted by the job manager. This results in some special
cases in the job manager that can be confusing.

Since job-ingest now waits for the job manager to "approve" each
job before responding to the user, this division of labor is no
longer required.

Have the job manager post the submit event.  Maintain the invariant
that the user does not receive the job ID until the submit event
has been committed to the KVS by deferring the response to job-ingest
until the KVS batch completes.

Impact on submit latency:  latency is increased because job-ingest does
not send its request to the job manager until the jobspec is committed
to the KVS, and now the job manager does not respond until the submit
event is committed to the KVS.  An individual submit must now wait for
two back to back KVS commits instead of one.

Impact on job throughput:  any impact on job throughput should be
mitigated by KVS commit batching.  For example, the submit event is
likely now batched with the depend event that usually follows it,
whereas before it was batched with the jobspec.