job-manager: drop code for restart race
Problem: when job-ingest posted the submit event, there could
be a race where a submit request could be received for a job
already in the hash because it was discovered during KVS restart.

The submit event is now posted by the job manager, so this condition
can no longer occur.

Drop workaround code and comments for this case.