kvs: refactor state machine from switch to loop
Problem: The current way that kvs transactions move
between states is to "fallthrough" to additional case
statements in a switch statement.  When new states
are added that can jump to different states, the
fallthrough mechanism will not work.

Solution: Convert the switch statement into a while
loop, so that state transitions occur on another
iteration of the loop.  Adjust some comments as a result.