kvs: initialize kvsroot mgr correctly
Problem: When creating the kvsroot manager, an unitialized value
of "ctx->h" is passed in, which sets a NULL handle in the kvsroot
manager.

Solution: Set ctx->h earlier to ensure ctx->h is set when being used.
For consistency, use "ctx->h" instead of "h" for the flux handle
in nearby calls.