libsubprocess: check for args in local exec
Problem: flux_local_exec() did not check if the user supplied
any arguments for execution.  This was not consistent to the
checks supplied by flux_rexec() and can lead to a segfault
internally.

Solution: Add a argument count check identical to the check done
in flux_rexec().  Add a unit test.

Fixes #4347