testsuite: fix jobtap event posting test
Problem: jobtap event posting test in t2212-job-manager-plugins.t
does not fail unless the last state in the for loop fails.

Refactor script so that each state is run in a separate test.
This also separates the outupt for each test, making it easier to read.
Expect failure for state=new,validate since those cannot emit events
in the current design.