flux-kvs: add --blobref options to kvs writes cmds
Problem: The addition of flux_kvs_commit_get_rootref()_makes it so
the blobref of a commit can be output after completion, but there no
kvs commands that support this yet.

Solution: Support a --blobref / -b option for the put, unlink, link,
and mkdir commands in flux-kvs.  Add sharness tests.