job-manager: replace priority-hold w/ submit-hold
Problem: the priority-hold jobtap plugin sets the job's priority to 0
to hold the job and expects that the job can be allowed to be scheduled
by manually updating the job's urgency, but now that the urgency event
transitions the job back to PRIORITY state, priority-hold resets the
priority to 0 again and the job cannot be released.

Change the plugin to immediately post an event that sets the urgency to
FLUX_JOB_URGENCY_HOLD (0) on entry into DEPEND state, overriding the
user's submit-time urgency.  As before, the job can be "unheld" by updating
the urgency.

Rename the plugin from 'priority-hold' to 'submit-hold' since it no
longer sets the priority directly.

Update jobtap sharness test.