job-manager: don't validate dependencies on posting
Problem: since dependency-add and -remove events are not
necessarily processed immediately, validating them against
the current grudgeset may result in inappropriate failures.

For example, a remove could look incorrect if the matching add
event was queued, since at the time the remove event is posted,
the dependency may not yet be in the grudgeset.

Relax checking of dependency-add and -remove events at the time
of posting.  Any errors will be deferred to when the events are
processed.

Update t2270-job-dependencies.t.