cmd: replace strcmp() in flux-job.c with streq()
Problem: flux-job makes heavy use of !strcmp() to compare if two
strings are equal, but this is much less readable than the streq()
macro made available via the internal CCAN library.

Switch use of !strcmp() to streq() in flux-job.c. Update the other
uses of strcmp() as well for consistency.