flux-job: add -w, --wait-event option to flux-job attach
Problem: flux-job attach will always wait for the final event in the
job eventlog (the 'clean' event), even though in most cases it will
be sufficient to wait for the finish event, along with completion of
all i/o.

Add a new option `-w, --wait-event=NAME` to allow this behavior to
be changed. For now, the behavior of waiting for the 'clean' event
is kept to avoid breaking existing tests.