testsuite: use flux-job attach --wait-event flag where necessary
Problem: A couple tests in the testsuite rely on the current behavior
of flux-job attach waiting for the clean event in the job eventlog.
Howevetr, this will break once the default wait-event for flux-job
attach is changed to the finish event.

Update t2500-job-attach.t to ensure that the finish event is seen by
default and the clean event is emitted with `--wait-event=clean`.

Update t2274-manager-perilog.t to force waiting for the clean event
to ensure that the epilog-start and epilog-finish events are seen
during the test.