job-manager: reject exceptions posted in NEW state
Problem: if a fatal exception is raised on a job in NEW state,
it transitions to INACTIVE, but this is not a legal state
transition per RFC 21.

Log an error message and return an error when the exception
is posted, since in NEW state it will not have the desired effect
and should be considered a coding error.

To make the intention crystal clear, change the state machine to
ignore fatal exceptions posted in NEW state even though that should
be prevented by the above change.