resource: open resource.topo-get RPC to users
Problem: The hwloc topology is already loaded by the resource module,
and is made available via the resource.topo-get RPC, but this RPC is
open to the instance owner only.

Open the RPC to FLUX_ROLE_USER so that guest users (such as the job
shell) can fetch the topology XML instead of recreating it.