broker: return error if no backing store
Problem: The content.flush target will never respond to a request
if a backing store is never loaded.

Solution: Return error with ENOSYS if content.flush cannot be
completed on rank 0.  In addition, return ENOSYS if a content.flush is in
progress when the backing store is unregistered.  Add regression test.

Fixes #4375