broker/content-cache: add entries to flush list
Problem: If the backing store has not been registered, dirty
cache entries will never be added to the flush list.  If a backing
store is ever added later, those missing dirty entries will never
be added to the backing store.

Solution: If the backing does not exist, still add dirty entries
to the flush list.  Note that dirty entries are not dropped from
the cache during a purge or dropcache, so this does not change
behavior.

Fixes #4379