broker/content-cache: do not remove flush entries
Problem: In cache_flush(), all entries on the 'flush' list are
popped off in the function loop.  This is regardless if any entries
lead to errors when calling cache_store().  Some errors, such as
ENOSYS or ENOMEM may be temporary, and we do not want those entries
lost on the flush list forever.

Solution: Break out of the loop if certain error conditions occur
like ENOSYS or ENOMEM, and we have confidence that continued
iteration in the loop would be fruitless.  Also, do not pop the
cache entry off of the flush list until we've confirmed that
cache_store() has succeeded.