kvs: remove excess logging of ENOTSUP
Problem: When a namespace is not available (ENOTSUP) the kvs
module logs an error.  However, a namespace not available
error can be quite common, such as when the job-info module
transitions reading from a guest namespace to a primary namespace.
These log messages can be excessive in the logs.

Solution: Remove log messages when ENOTSUP is returned to
callers.  Let callers decide if ENOTSUP should be logged or not.

Fixes #4369