python: add LSF URI resolver plugin
Problem: In order for users to connect to Flux sessions on LSF-based clusters,
they currently have to go through a series of string parsing and shell calls
to get the URI of the session and connect via flux proxy. With replicating these
steps, users now have a way to connect to sessions using just the job ID of their
LSF job, as long as a flux session is running on all the compute nodes in that job.

Fixes flux-framework/flux-core#4377