job-manager: perilog: allow job manager epilog without a prolog
Problem: The call to flux_jobtap_job_subscribe(3) in the perilog.so
job manager plugin is made during the run action of the job manager
prolog. This means that an epilog action cannot run unless a prolog
is also registered.

Move the call to subscribe to job events into the job.state.run
callback directly. Only register for events if there is a perilog
action configured.