job-manager: add jobspec-default plugin
Problem: there is no mechanism to configure default values for
jobspec system attributes that are left unspecified by the user.

Add a builtin jobtap plugin that parses the
[policy.jobspec.defaults.system] TOML table described in RFC 33.
If a key is present in the table but missing from the jobspec
system attribute object, a jobspec-update event is posted, which
updates the job manager's in-memory jobspec for the job.

If a queue is assigned, the queue specific default policy is then
applied, overriding any duplication in the general defaults, as
described RFC 33.

Note that due to the way duration was specified in RFC 14, we treat
a duration of zero in the jobspec the same as if it were missing.