kvs: add easy kvsroot check for primary namespace
Problem: Checking for the special case that a namespace
is the primary namespace can be costly because it constantly
involves doing an strcmp() on a string.

Solution: Add a special flag to the kvsroot structure that
does a onetime check if the root happens to be the primary
namespace root.