job-manager: support after* dependencies on inactive jobs
Problem: Job manager plugins can now lookup inactive jobs via the
flux_jobtap_job_lookup(3) function, but the dependency-after plugin
does not properly handle jobs in the INACTIVE state.

Add a check for a job state of INACTIVE in the dependency-after
plugin, and attempt ot either resolve the depedency immediately or
reject the job if the dependency is not satisfiable.

Fixes #4287