subprocess: improve efficiency of fd closing in child process
Problem: When closing file descriptors in the subprocess child, each
examined file descriptor must iterate the hash of channels to check
if the current file descriptor should remain open in the child. This
is inefficient, especially when there are many open file descriptors
passed in from the parent process.

Add a function that returns an idset containing all child file
descriptors and use this in closefd_child() to test if the currently
examined fd belongs to the child process.