flux-exec: allow --dir=none to unset current working directory
Problem: flux-exec always sets cwd in the flux_cmd_t passed
to libsubprocess, so there is no way to take advantage of the
posix_spawn() implementation from this utility.

Support a special "none" argument to unset the current working dir
for the remote subprocess. The subprocess will simply inherit the
current working directory of the broker.

This is mostly meant for testing the posix_spawn() implementation
from the flux-exec(1) front end utility.