subprocess/test: fix exec error tests on some systems
Problem: On some versions of glibc, posix_spawnp(3) is not able to
return appropriate errors to the caller for certain exec errors
(e.g. EACCES). This causes a couple tests in the subprocess unit
tests to fail.

Modify the failing tests to use a current working directory so that
we fall back to using fork/exec (the implementation the tests are
meant for anwyay)