shell: set working directory before loading initrc
Problem: The shell is no longer invoked in the job working directory by
the job-exec system, and currently the shell initrc is loaded before
the shell itself changes its working directory. This can lead to a
failure to load an initirc specified by a relative path at submission
time where this worked before.

Move the load of the initrc to after the shell changes its working
directory so that relative initrc paths continue to work.