testsuite: change working directory in dummy shell
Problem: Since the job execution system no longer changes the working
directory of the job shell before exec, the dummy job shell in the
testsuite may not be able properly execute relative path arguments
because it is in the wrong directory.

Change the directory of the dummy job shell before executing its
arguments to fix the issue.