job-manager: avoid job aux destroy segfault
Problem: plugins may store data in the job aux container with
a destructor implemented within the plugin.  If the plugin is
unloaded, the broker could segfault when the aux hash is destroyed
later on, if the destructor is no longer resident or has moved.

Instead of storing items directly in the job aux container, store
a plugin-specific container within the job aux container.  As the
plugin is being unloaded, the outer container is destroyed before the
plugin is destroyed.  In addition, the outer container destructor
ensures that the plugin is still loaded before destroying the inner
container.