job-manager: clear aux container on inactive jobs
Problem: plugins may store arbitrary per-data in the job's
aux container, and now that inactive jobs are retained,
these items may remain in memory for a long time and for no
purpose if they are not explicitly cleaned up by the plugin.

Clear the aux container when a job transitions to inactive.
Leave a note in job.h for future users outside of jobtap.

Fixes #4296