job-manager: handle pre-0.41 job eventlogs
Problem: flux-core 0.41.0 refuses to start with older
jobs in the KVS.

The state machine changed in 0.41 to require a "validate" event
to transtion the job from NEW to DEPEND, while before, the "submit"
event did that.

To address this, RFC 21 now requires a version attribute in the
(always first) submit event.

For now, add a work around so that if the version attribute is missing,
the eventlog is assumed to be pre-0.41.1 and the job state is advanced
to DEPEND after the submit event is replayed.

Update unit tests.

Fixes #4398