ci: generate dumpfile from most recent tag for testing
Problem: There is no automated way to ensure that a PR does not break
reloading jobs from the most recent tag of flux-core. This could lead
to losing all jobs on the next restart after an upgrade.

Add a script which can generate a dumpfile from the most recent tag of
flux-core by using the official tagged docker image. Run this script
in CI and copy the result to t/job-manager/dumps/valid to ensure this
dumpfile can be read by the currently tested PR.

Fixes #4401