t2802-uri-cmd.t: add test for LSF resolver
Problem: The addition of an option to resolve Flux URIs using LSF job ID
requires additions to the flux-core testsuite.

Add a test that tests the LSF URI resolver using a fake ssh and csm_
allocation_query scripts. Previous commit updated to detect SSH path
for integration with this test.