job-manager: add limit-job-size plugin
Problem: flux lacks the ability to configure simple limits
on job size.

Add a jobtap builtin plugin "limit-job-size" that allows the
simple job size limits described in RFC 33 to be implemented.

General limits may be set in the TOML [policy.limits] table.
These limits may be overridden for a given queue in the
[queues.<name>.policy.limits] table.