testuite: add test for eventlog race with job start
Problem: There is no regression test for #4409, where the commit of
the job eventlog in the job manager races with job startup and the
job shell's attempt to watch job eventlogs.

Use the new job-manager.set-event-timeout RPC to set a very long
event batch timeout in the job manager. Then run a series of jobs
and ensure they all complete as expected, as opposed to throwing job
exceptions due to missing eventlogs.