job-info: fix empty eventlog corner case
Problem: If an eventlog is empty (which is a RFC18 violation),
a ENOMEM error is returned from a job-info lookup due to a NULL
pointer being passed to json_string().

Solution: Check for empty job values and return EPROTO.  While it
is feasible that a job value could have an empty value, all known
major ones (eventlog, jobspec, R, etc.) are required to be non
empty.

Fixes #4413