job-info: return EPROTO on invalid eventlog
Problem: In several circumstances that an eventlog is invalid,
EINVAL is returned to the caller.  But this is not consistent to
some other changes.  In addition, as RFC18 and RFC21 clearly define
what the eventlog should look like, EINVAL seems like the wrong
error value.

Solution: Return EPROTO instead of EINVAL when the eventlog is invalid
in some way.