flux-dump: fix error on blobref with empty value
Problem: flux-dump fails with an error such as

 flux-dump: key: missing blobref N: Protocol error

when the associated blobref is an empty value. This occurs because
of flux-dump's use of flux_msg_get_payload(3), which returns EPROTO
when the payload is empty, because FLUX_MSGFLAG_PAYLOAD is never set
for messages with a zero length payload.

Switch flux-dump to use flux_response_decode_raw(3), which handles
the empty payload by setting len=0 and buf=NULL. Then, only call
dump_write_data() if len > 0 to ensure empty values are preserved
as empty.

Fixes #4414