librlist: preserve node up/down state in rlist_copy_match()
Problem: When copying an rlist by matching constraints, the state of
the nodes copied is not preserved, and therefore all nodes are copied
as up, which could cause the scheduler to allocate down nodes to a job.

Copy the up/down state of the rnode in rnode_copy_match(), so that the
resulting rlist from rlist_copy_constraint() contains the state of the
copied nodes.

Fixes #4424