testsuite: enhance sched-simple constraint matching tests
Problem: No tests in the testsuite ensure that sched-simple will not
allocate a down node to a job which uses the --requires flux-mini option.
Also, tests for feasibility checks are missing from the paltry existing
tests in t2276-job-requires.t.

Enhance the tests in the sched-simple requires testing to include missing
regression and functionality tests.