kvs: cleanup by using bools instead of bitmask
Problem: In the kvstxn code, an internal bitmask of flags is used
for "internal" purposes.  The use of a bitmask for internal purposes
can be confused with other flag parameters that may be passed in
by users.

Solution: Remove the "internal_flags" and replace with bools representing
the conditions the internal code cares about.