t2276-job-requires.t: add noverify to resource module reload
Problem: When t2276-job-requires.t reloads the resource module, it
does not provide the noverify option. This could lead to some ranks
being drained due to resource mismatch since the test is using fake
resources (2 cores per node).

Add the "noverify" option when reloading the resource module in
this test.