flux-top: avoid premature exit when recursive top fails
Problem: A comment in joblist_pane_enter() indicates that a failure
to recursively enter a job should cause top to remain in the original
window.  However, flux-top currently exits if it cannot attach to
the new job.

Remove calls to fatal() from top_run() and instead return an error
string if the function fails. Print this error string in main()
to preserve current behavior, but ignore the error in the recursive
joblist_pane_enter().