content-sqlite,files,s3: rename checkpoint topics
Problem: In the near future we would like the KVS checkpoint
to content modules to be routed through the broker, instead
of communicating directly to the modules.  The topic names
of kvs-checkpoint.get and kvs-checkpoint.put should be renamed
similarly to other topics in the content modules so readers
will know they all behave in similar fashion.

Solution: Rename kvs-checkpoint.get and kvs-checkpoint.put to
content-backing.checkpoint-get and content-backing.checkpoint-put
respectively.  Update all callers and tests accordingly.  Due to
the length of the topic name, adjust some text from
"content-backing.checkpoint" to just "checkpoint" when appropriate.