broker: add content.checkpoint-{get,put} targets
Problem: In the future, we would like to cache checkpoint
data in the broker in the event a content-backing module is
not loaded.  But this requires checkpoint data to be routed
through the broker and not sent directly to the content-backing
modules.

Solution: Add new content.checkpoint-get and content-checkpoint-put
handlers in the broker which will route to content-backing.checkpoint-get
and content-backing.checkpoint-put if they are available.  Update
callers to call content.checkpoint-{get,put}.

Update t0018-content-files.t tests that had previously assumed no
registration with the broker.

Update t2807-dump-cmd.t test that will get different error message
as a result.