job-manager: recover max jobid during KVS replay
Problem: the KVS checkpoint.job_manager key that contains the max
job ID is only written out when the job manager is unloaded.  If the
unload fails, or if the broker crashes before then, an older value
may be used and job IDs could be reissued.

Track the max job id encountered during KVS job replay, and use the
greater of the replayed job IDs or the saved max job ID value.

Fixes #4300