libccan: add array_size module
Problem: we use "sizeof (a) / sizeof (a[0])" a lot to determine
array size, which is rather verbose and usually involves defining
the size as a separate const variable to avoid using the expression
in the wrong context.

The ccan ARRAY_SIZE() macro is equivalent to the above and includes
protection against use in the wrong context.

Pull it in to our copy of libbcan.