libutil: popen2: return wait status in pclose2()
Problem: pclose2() returns -1 with EIO if the command returns a nonzero
wait status, but this could be confusing since the standard pclose(3)
returns the wait status of the command, even if it is nonzero.

If the command exited, return the wait status in pclose2() instead of
-1 with errno=EIO.

Update one affected unit test.