libflux: add flux_open_ex(3)
Problem: flux_open(3) may emit errors to stderr, especially in the
case of the ssh connector, which spawns a child process with popen2().
This is not considered good behavior from a library, and can cause
problems for programs using libflux which do not expect to share
their stderr with the library.

Add a flux_error_t parameter to the internal connector_init()
function used to initialize flux handles. In the ssh connector, if
this parameter is non-NULL direct errors to the error buffer instead
of stderr through use of popen2() capture stderr flag and other means.

Add a new flux_open() interface flux_open_ex() which takes an optional
flux_error_t parameter, which is then passed to connector_init(),
allowing callers access to errors outside of stderr.

Also update find_connector() to take a flux_error_t so that more
descriptive errors can be returned, since we already have access to
the flux_error_t parameter, when provided.