flux-top: popup error dialog on recursive top failure
Problem: If top_create() on a child job in the joblist pane fails,
then there is no error returned to the user.

Capture the error from top_create() and display it in an error dialog
in the joblist pane. The dialog will go away in 4 seconds by default,
or may be dismissed by pressing any key.