testsuite: test flux_open_ex(3) capture of ssh stderr
Problem: No tests in the testsuite verify that the stderr of ssh
in the ssh connector is redirected away from the current processs'
stderr, and is captured in the error parameter.

Utilize the fact that the Python flux.Flux() constructor now uses
flux_open_ex(3) to test that ssh stderr is captured.