flux-relay: initialize log prefix to hostname when possible
Problem: Users almost always use flux-relay indirectly via the ssh
connector by specifying a ssh:// URI or a high-level URI which resolves
to an ssh URI. Therefore, the flux-relay log prefix on errors from
this utility may not be useful to users.

If possible, use the hostname as the log prefix. This will not only
remove the possibly confusing flux-relay prefix, but will allow
pinpointing the actual host on which the flux-relay is failing to,
for example, open a local connector socket. In the unlikely event that
gethostname(3) fails, then fall back to "uknown-host".