content-sqlite,files,s3: register module later
Problem: In the content-sqlite, content-files, and content-s3 modules,
the backing module is registered before setup of module services is
complete.  This can cause problems if the broker's content-cache
wants to make calls into the module during registration time.

Solution: Ensure that all setup is complete before registering the
backing module.

Update regression test which would fail with this change.

Fixes #4456