flux-overlay: fix incorrect disconnect message
Problem: flux overlay disconnect prints the same hostname
for parent and target, e.g.
  flux-overlay: asking picl2 (rank 0) to disconnect child picl2 (rank 2)

This is because flux_get_hostbyrank () clobbers the previous value
and we call it twice to construct the message.

Make a copy of one values before the next call, then free the copy
after printing the message.