github: only copy test dumpfile if one was generated
Problem: In the "generate dumpfile from most recent tag" step, a
dumpfile is not generated when building from a tag, since the most
recent tag is the current tag, but the step tries to copy the
dumpfile anyway and then fails.

Only copy the dumpfile if it was created. If the generation of the
dumpfile fails for other reasons, then the script will have failed,
so the step will still fail CI.