github: switch to maintained release action
Problem: The GitHub releases action is no longer maintained, is split
into two different steps in order to upload an asset, and also seems to
trigger a release when a user pushes to their own fork of the project.

Switch to a maintained release action, softprops/action-gh-release@v1,
which has the side benefit of supporting files to updload as assets
for the release. Add github.repository == 'flux-framework/flux-core' to
ensure auto-deployed releases only occur in the main flux-core project.