broker: update cache entry removal accounting
Problem: In cache_entry_remove(), the `acct_dirty` variable is
decremented if the entry is dirty.  However, cache_entry_remove()
should never be called if a cache entry is dirty.

Solution: Remove the decrement of `acct_dirty` if a cache entry
is dirty.  Instead add an assert() check that dirty is false.