broker: use topology class
Problem: the broker overlay code uses the kary class to calculate
peers and routes, but this will not work once the topology is allowed
to have a non-complete tree shape.

Switch to the topology class, which supports any tree shape.  The topology
is created in the bootstrap methods and passed to the overlay using
overlay_set_topology(), which replaces overlay_set_geometry().

Use the topology_set_kary() method to enact a complete topology as
before, but never call the kary_*() functions directly and remove
all assuptions about complete trees from the code.

Note that formerly the fact that child peers had contiguous ranks
was used to directly index into the array of children when selecting
a route.  To avoid slowing down that critical path, use the topology
aux container to store a pointer to the child peer struct.  The
container is accessed in O(1), and its items in O(N) where N is the
item's position in the container.

Update the overlay unit test.