broker: move tbon.fanout to bootstrap methods
Problem: in the future, the broker bootstrap methods (config file
and PMI) will handle the tbon.fanout broker attribute differently,
but the attribute is currently managed in shared code in overlay.c.

Move the handling of the attribute to the bootstrap methods where
the code can be allowed to diverge.

Drop tbon.fanout test from overlay unit test, since this is no longer
a feature that belongs to the overlay abstraction.

Allow a fanout value of 0 to indicate "flat", regardless of the
instance size, since the topology class supports that.  Update
t0001-basic.t sharness test to allow that value to succeed.
Update flux-broker-attributes(7) to say what a value of 0 means.