flux-mini: combine --setattr/--setopt code
Problem: The flux mini --setattr and --setopt options are handled
differently, which seems unnecessary and could be confusing to users.

Add a new function parse_jobspec_keyval() which combines common code
for handling these options which take jobspec key=value arguments.

As a side effect, --setattr now defaults the value of keys to 1 if
no value is specified, and --setopt can now read values from a file
if the key is prefixed with "^".

Update one test that expects --setattr=key with no value to fail.