flux-mini: use attributes.system as default for --setattr
Problem: The flux-mini --setattr option is almost always used to set
attributes in the system dictionary, and therefore its use can be
very repetitive and inefficient.

Prepend "system." to the KEY specified in --setattar=KEY=VAL when
it doesn't already begin with "system." or "user.". To force
a different top-level target value, also allow keys to be prefixed
with "." (e.g. --setattr=.foo will set attributes.foo=1).

Fix one use of --setattr=foo=val in the testsuite with
--setattr=.foo=val.

Fixes #4481