t2700-mini-cmd.t: add more tests for flux-mini --setattr
Problem: The flux-mini --setattr tests do not verify that --setattr=KEY
sets the attribute KEY in the system dictionary by default, and that
--setattr now defaults the value to 1 when value is not specified.

Test these conditions in the existing --setattr tests in
t2700-mini-cmd.t.

Also, update the test use jq directly to compare values instead of
reading the values with jq and comparing with test.