broker: avoid double adding cache entry onto list
Problem: A dirty cache entry has to potential to be added onto the
content cache's flush list twice.  This double addition can lead to list
corruption.  The observed side effect was a list that was shortened and no
longer accurate with respect to the `acct_dirty` counter.  This could lead
to hangs with content flush, missed flushes to the backing store, and
segfault/memory corruption in the worst case.

Solution: Remove the cache entry from the flush list before adding it.
The remove is a no-op if it is not already on a list.

Fixes #4482